# Reverse lookup for DataFlowAnalyzer