--- a/libyul/CMakeLists.txt
+++ b/libyul/CMakeLists.txt
@@ -183,6 +183,8 @@ add_library(yul
 	optimiser/UnusedPruner.h
 	optimiser/VarDeclInitializer.cpp
 	optimiser/VarDeclInitializer.h
+	optimiser/VariableAssignmentMap.cpp
+	optimiser/VariableAssignmentMap.h
 	optimiser/VarNameCleaner.cpp
 	optimiser/VarNameCleaner.h
 )