@stackenbotten 2023-04-12
There was an error when running `chk_coding_style` for commit `830857aa59c005a43fa63a9e9d9a4414f019202b`:
```
Coding style error:
libyul/optimiser/VariableAssignmentMap.cpp:17:    for (auto&& reference: m_ordered[_variable])
libyul/optimiser/VariableAssignmentMap.cpp:18:        if (m_reversed.find(reference) != m_reversed.end())
libyul/optimiser/VariableAssignmentMap.cpp:19:        {
libyul/optimiser/VariableAssignmentMap.cpp:20:            if (m_reversed[reference].size() > 1)
libyul/optimiser/VariableAssignmentMap.cpp:21:                m_reversed[reference].erase(_variable);
libyul/optimiser/VariableAssignmentMap.cpp:22:            else
libyul/optimiser/VariableAssignmentMap.cpp:25:                m_reversed.erase(reference);
libyul/optimiser/VariableAssignmentMap.cpp:26:        }
libyul/optimiser/VariableAssignmentMap.cpp:27:    m_ordered.erase(_variable);

```
Please check that your changes are working as intended.