@cameel 2023-05-26
I checked how this PR affects compilation times in external tests.
Here are the times for `ir-optimize-evm+yul` preset from [run 29920](https://app.circleci.com/pipelines/github/ethereum/solidity/29920/workflows/c61d231c-40a5-4415-8edc-bef92ce843dd) on `develop` (8c7404f639b19a30d0e0ace31002bb48a7a6f5c4) vs [run 29928](https://app.circleci.com/pipelines/github/ethereum/solidity/29928/workflows/7d567d99-859b-4c91-aa81-07a5214f7a91) on `dataflow-analyzer-reverse-lookup` (a8cc9bde9b64d44e19b59ada18412675654e5d22).

| test             | `develop` | this PR   | difference |
|------------------|-----------|-----------|------------|
| zeppelin         | 6m48.256s | 5m46.622s | -1:02      |
| colony           | 1m51.012s | 1m27.582s | -0:24      |
| uniswap          | 1m46.737s | 1m40.339s | -0:06      |
| prb-math         | 0m13.031s | 0m10.887s | -0:03      |
| gp2              | 0m44.495s | 0m43.621s | -0:01      |
| brink            | 0m04.842s | 0m05.302s | +0:01      |
| ens              | 0m45.856s | 0m48.436s | +0:03      |
| perpetual-pools  | 1m28.025s | 1m32.749s | +0:04      |
| yield-liquidator | 0m31.450s | 0m38.634s | +0:07      |
| elementfi        | 2m48.506s | 3m15.299s | +0:27      |

I wonder what's happening with ElementFi. I double checked I got the numbers from the right pages and looks like it really took ~30 s longer. Maybe worth checking if it's repeatable or just a fluke. Still, some other times increased too so there might be something more to it.

The table contains only the `real` time. Full results are below in case anyone's interested.
<details>

```
zeppelin
real	6m48.256s user	6m47.482s sys	0m2.189s
real	5m46.622s user	5m46.191s sys	0m1.551s

uniswap
real	1m46.737s user	1m48.668s sys	0m0.732s
real	1m40.339s user	1m41.697s sys	0m0.673s

ens
real	0m45.856s user	0m45.927s sys	0m0.656s
real	0m48.436s user	0m48.450s sys	0m0.580s

gp2
real	0m44.495s user	0m52.945s sys	0m1.442s
real	0m43.621s user	0m51.973s sys	0m1.369s

brink
real	0m4.842s user	0m4.763s sys	0m0.603
real	0m5.302s user	0m5.272s sys	0m0.556s

elementfi
real	2m48.506s user	2m56.110s sys	0m2.225s
real	3m15.299s user	3m24.894s sys	0m2.584s

perpetual-pools
real	1m28.025s user	1m30.875s sys	0m1.107s
real	1m32.749s user	1m36.234s sys	0m1.284s

prb-math
real	0m13.031s user	0m13.804s sys	0m0.469s
real	0m10.887s user	0m11.300s sys	0m0.444s

yield-liquidator
real	0m31.450s user	0m34.304s sys	0m0.567s
real	0m38.634s user	0m41.636s sys	0m0.534s

colony
real	1m51.012s user	2m17.810s sys	0m3.484s
real	1m27.582s user	1m50.964s sys	0m2.876s
```

</details>