@cameel 2024-04-19
I gathered fresh timing data (and translated the original data into the same format).

1. It seems to me that all the timing differences we see in external tests, even the positive ones may be just noise:
    - The differences are much smaller now. Especially for `zeppelin`, `elementfi` and `colony`, which originally had the largest difference, the difference not only decreased but completely reversed direction.
    - In some cases the differences flip between positive and negative in different runs for the same project.
    - `zeppelin` and `uniswap` compile much faster on `develop` now. In case of `zeppelin` this could very well be due to upstream changes in tests. For `uniswap` though we have a forked repo.
2. On the other hand, there is actually a 2x difference in our own timing benchmark - though only for a single contract (`chains.sol`). Others, e.g. `OptimizorClub.sol`, do not seem affected.

My conclusion here is that there's no strong evidence that this change significantly affects external tests, be it positively or negatively, but it does seem to help one especially pathological contract from our repo (`chains.sol`). I wonder if it might be because profiling was done specifically on that contract and the PR addresses its specific bottleneck, which happens to be different than in more normal scenarios?

### Original timing
| Project          | `develop` | this PR |
|:-----------------|----------:|--------:|
| brink            |       5 s |     5 s |
| colony           |     111 s |    88 s |
| elementfi        |     169 s |   195 s |
| ens              |      46 s |    48 s |
| euler            |           |         |
| gnosis           |           |         |
| gp2              |      44 s |    44 s |
| perpetual-pools  |      88 s |    93 s |
| pool-together    |           |         |
| uniswap          |     107 s |   100 s |
| yield_liquidator |      31 s |    39 s |
| zeppelin         |     408 s |   347 s |
| prb-math         |      13 s |    11 s |

### Current timing
| Project          |`develop` | this PR (run 1) | this PR (run 2) | this PR (run 3) |
|:-----------------|---------:|----------------:|----------------:|----------------:|
| brink            |      5 s |             4 s |             5 s |             5 s |
| colony           |    105 s |           113 s |           110 s |           114 s |
| elementfi        |    172 s |           162 s |           163 s |           167 s |
| ens              |     40 s |            43 s |            43 s |            47 s |
| euler            |     52 s |            59 s |            62 s |            72 s |
| gnosis           |          |                 |                 |                 |
| gp2              |     42 s |            48 s |            47 s |            60 s |
| perpetual-pools  |     81 s |            72 s |            84 s |            79 s |
| pool-together    |     56 s |            49 s |            51 s |            61 s |
| uniswap          |     82 s |            84 s |            71 s |            84 s |
| yield_liquidator |     26 s |            27 s |            26 s |            25 s |
| zeppelin         |    261 s |           266 s |           274 s |           277 s |
| prb-math         |          |                 |                 |                 |

- [run on `develop`](https://app.circleci.com/pipelines/github/ethereum/solidity/33792/workflows/82fd7084-3a37-414f-abc0-bcfd26f76ec1)
- [run 1 on `dataflow-analyzer-reverse-lookup`](https://app.circleci.com/pipelines/github/ethereum/solidity/33793/workflows/040c388e-9515-4d21-b70e-73c190d10a39)
- [run 2 on `dataflow-analyzer-reverse-lookup`](https://app.circleci.com/pipelines/github/ethereum/solidity/33793/workflows/6d4d754d-9fd2-49b7-90d2-9a2854600282)
- [run 3 on `dataflow-analyzer-reverse-lookup`](https://app.circleci.com/pipelines/github/ethereum/solidity/33793/workflows/598bad99-95f0-42de-acbd-4dab7b8aaac3)

### Timing diff
| Project          | Diff (original) | Diff (run 1) | Diff (run 2) | Diff (run 3) |
|:-----------------|----------------:|-------------:|-------------:|-------------:|
| brink            |             1 s |         -1 s |          0 s |          0 s |
| colony           |           -24 s |          8 s |          5 s |          9 s |
| elementfi        |            27 s |        -10 s |         -9 s |         -5 s |
| ens              |             3 s |          3 s |          3 s |          7 s |
| euler            |                 |          7 s |         10 s |         20 s |
| gnosis           |                 |              |              |              |
| gp2              |            -1 s |          6 s |          5 s |         18 s |
| perpetual-pools  |             4 s |         -9 s |          3 s |         -2 s |
| pool-together    |                 |         -7 s |         -5 s |          5 s |
| uniswap          |            -6 s |          2 s |        -11 s |          2 s |
| yield_liquidator |             7 s |          1 s |          0 s |         -1 s |
| zeppelin         |           -62 s |          5 s |         13 s |         16 s |
| prb-math         |            -3 s |              |              |              |

### Timing benchmark
#### `develop`
Binary from `b_ubu_static` from [the last run on `develop`](https://app.circleci.com/pipelines/github/ethereum/solidity/33792/workflows/82fd7084-3a37-414f-abc0-bcfd26f76ec1)
| File                 | Pipeline | Bytecode size | Time     | Exit code |
|----------------------|----------|--------------:|---------:|----------:|
| `verifier.sol`       | legacy   |    4940 bytes |   0.20 s |         0 |
| `verifier.sol`       | via-ir   |    4417 bytes |   0.66 s |         0 |
| `OptimizorClub.sol`  | legacy   |       0 bytes |   0.52 s |         1 |
| `OptimizorClub.sol`  | via-ir   |   22391 bytes |   4.05 s |         0 |
| `chains.sol`         | legacy   |    5878 bytes |   0.17 s |         0 |
| `chains.sol`         | via-ir   |   23076 bytes |  23.16 s |         0 |

#### this PR
Binary from `b_ubu_static` from [the last run on `dataflow-analyzer-reverse-lookup`](https://app.circleci.com/pipelines/github/ethereum/solidity/33793/workflows/598bad99-95f0-42de-acbd-4dab7b8aaac3)

| File                 | Pipeline | Bytecode size | Time     | Exit code |
|----------------------|----------|--------------:|---------:|----------:|
| `verifier.sol`       | legacy   |    4940 bytes |   0.14 s |         0 |
| `verifier.sol`       | via-ir   |    4417 bytes |   0.65 s |         0 |
| `OptimizorClub.sol`  | legacy   |       0 bytes |   0.57 s |         1 |
| `OptimizorClub.sol`  | via-ir   |   22391 bytes |   4.23 s |         0 |
| `chains.sol`         | legacy   |    5878 bytes |   0.17 s |         0 |
| `chains.sol`         | via-ir   |   23076 bytes |  10.52 s |         0 |